Separated rendering into base class
