+++ addons/verlet_rope_4/VerletRope.cs
@@ -1,43 +1,12 @@
 using Godot;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using VerletRope.addons.verlet_rope_4.Structure;
 using VerletRope4.Structure;
 
 namespace VerletRope4;
 
-/*
-MIT License
-
-Rewritten to be used with Godot 4.0+.
-No additional license applied, feel free to use without my notice, though do not forget to still apply original license.
-(c) Timofey Ivanov / tshmofen
-
-Copyright (c) 2023 Zae Chao(zaevi)
-Copyright (c) 2021 Shashank C
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-*/
-
 [Tool]
-public partial class VerletRope : MeshInstance3D
+public partial class VerletRope : VerletRopeMesh
 {
     #region Signals
 
@@ -49,28 +18,18 @@ public partial class VerletRope : MeshInstance3D
 
     #region Vars Private
 
-    private const string DefaultMaterialPath = "res://addons/verlet_rope_4/materials/rope_default.material";
-    private const string NoNotifierWarning = $"Consider checking '{nameof(UseVisibleOnScreenNotifier)}' to disable rope visuals when it's not on screen for increased performance.";
-    private const string CreationStampMeta = "creation_stamp";
     private const string ParticlesRangeHint = "3,300";
-    private const string SimulationRangeHint = "30,265";
+    private const string SimulationRangeHint = "0,1000";
     private const string MaxSegmentStretchRangeHint = "1,20";
     private const string MaxCollisionsRangeHint = "1,256";
     private const float CollisionCheckLength = 0.001f;
     private const float DynamicCollisionCheckLength = 0.08f;
 
-    private static readonly float Cos5Deg = Mathf.Cos(Mathf.DegToRad(5.0f));
-    private static readonly float Cos15Deg = Mathf.Cos(Mathf.DegToRad(15.0f));
-    private static readonly float Cos30Deg = Mathf.Cos(Mathf.DegToRad(30.0f));
-
     private double _time;
-    private Camera3D _camera;
     private double _simulationDelta;
     private RopeParticleData _particleData;
-    private VisibleOnScreenNotifier3D _visibleNotifier;
 
     private RayCast3D _rayCast;
-    private ImmediateMesh _mesh;
     private BoxShape3D _collisionCheckBox;
     private PhysicsDirectSpaceState3D _spaceState;
     private PhysicsShapeQueryParameters3D _collisionShapeParameters;
@@ -113,9 +72,6 @@ public partial class VerletRope : MeshInstance3D
         get => _attachEnd;
     }
 
-    [Export] public float RopeLength { get; set; } = 5.0f;
-    [Export] public float RopeWidth { get; set; } = 0.07f;
-
     [Export(PropertyHint.Range, ParticlesRangeHint)]
     public int SimulationParticles
     {
@@ -134,19 +90,12 @@ public partial class VerletRope : MeshInstance3D
         get => _simulationParticles;
     }
 
-    private bool _useVisibleOnScreenNotifier = true;
-    [Export] public bool UseVisibleOnScreenNotifier
-    {
-        get => _useVisibleOnScreenNotifier; 
-        set { _useVisibleOnScreenNotifier = value; UpdateConfigurationWarnings(); }
-    }
-
     #endregion
 
     #region Vars Simulation
 
     [ExportGroup("Simulation")]
-    [Export(PropertyHint.Range, SimulationRangeHint)] public int SimulationRate { get; set; } = 60;
+    [Export(PropertyHint.Range, SimulationRangeHint)] public int SimulationRate { get; set; } = 0;
     [Export] public int Iterations { get; set; } = 2;
     [Export] public int PreprocessIterations { get; set; } = 5;
     [Export] public float PreprocessDelta { get; set; } = 0.016f; 
@@ -155,7 +104,6 @@ public partial class VerletRope : MeshInstance3D
     [Export] public bool Simulate { get; set; } = true;
     [Export] public bool Draw { get; set; } = true;
     [Export] public bool StartDrawSimulationOnStart { get; set; } = true;
-    [Export] public float SubdivisionLodDistance { get; set; } = 15.0f;
 
     #endregion
 
@@ -233,33 +181,9 @@ public partial class VerletRope : MeshInstance3D
 
     #region Util
 
-    private (Vector3 p0, Vector3 p1, Vector3 p2, Vector3 p3) GetSimulationParticles(int index)
-    {
-        var p0 = (index == 0)
-            ? _particleData[index].PositionCurrent - (_particleData[index].Tangent * GetAverageSegmentLength())
-            : _particleData[index - 1].PositionCurrent;
-
-        var p1 = _particleData[index].PositionCurrent;
-
-        var p2 = _particleData[index + 1].PositionCurrent;
-
-        var p3 = index == SimulationParticles - 2
-            ? _particleData[index + 1].PositionCurrent + (_particleData[index + 1].Tangent * GetAverageSegmentLength())
-            : _particleData[index + 2].PositionCurrent;
-
-        return (p0, p1, p2, p3);
-    }
-
     private float GetAverageSegmentLength()
     {
         return RopeLength / (SimulationParticles - 1);
-    }
-
-    private void ResetRopeRotation()
-    {
-        // NOTE: rope doesn't draw from origin to attach_end_to correctly if rotated
-        // calling to_local() in the drawing code is too slow
-        GlobalTransform = new Transform3D(Basis.Identity, GlobalPosition);
     }
 
     private float GetCurrentRopeLength()
@@ -290,87 +214,6 @@ public partial class VerletRope : MeshInstance3D
         collision = _rayCast.GetCollisionPoint();
         normal = _rayCast.GetCollisionNormal();
         return true;
-    }
-
-    #endregion
-
-    #region Draw
-
-    private static void CatmullInterpolate(Vector3 p0, Vector3 p1, Vector3 p2, Vector3 p3, float tension, float t, out Vector3 point, out Vector3 tangent)
-    {
-        // this is fast catmull spline
-
-        var tSqr = t * t;
-        var tCube = tSqr * t;
-
-        var m1 = (1f - tension) / 2f * (p2 - p0);
-        var m2 = (1f - tension) / 2f * (p3 - p1);
-
-        var a = (2f * (p1 - p2)) + m1 + m2;
-        var b = (-3f * (p1 - p2)) - (2f * m1) - m2;
-
-        point = (a * tCube) + (b * tSqr) + (m1 * t) + p1;
-        tangent = ((3f * a * tSqr) + (2f * b * t) + m1).Normalized();
-    }
-
-    private void DrawQuad(IReadOnlyList<Vector3> vertices, Vector3 normal, float uvx0, float uvx1)
-    {
-        // NOTE: still may need tangents setup for normal mapping, not tested
-        // SetTangent(new Plane(-t, 0.0f));
-        _mesh.SurfaceSetNormal(normal);
-        _mesh.SurfaceSetUV(new Vector2(uvx0, 0.0f));
-        _mesh.SurfaceAddVertex(vertices[0]);
-        _mesh.SurfaceSetUV(new Vector2(uvx1, 0.0f));
-        _mesh.SurfaceAddVertex(vertices[1]);
-        _mesh.SurfaceSetUV(new Vector2(uvx1, 1.0f));
-        _mesh.SurfaceAddVertex(vertices[2]);
-        _mesh.SurfaceSetUV(new Vector2(uvx0, 0.0f));
-        _mesh.SurfaceAddVertex(vertices[0]);
-        _mesh.SurfaceSetUV(new Vector2(uvx1, 1.0f));
-        _mesh.SurfaceAddVertex(vertices[2]);
-        _mesh.SurfaceSetUV(new Vector2(uvx0, 1.0f));
-        _mesh.SurfaceAddVertex(vertices[3]);
-    }
-
-    private float GetDrawSubdivisionStep(Vector3 cameraPosition, int particleIndex)
-    {
-        var camDistParticle = cameraPosition - _particleData[particleIndex].PositionCurrent;
-
-        if (camDistParticle.LengthSquared() > SubdivisionLodDistance * SubdivisionLodDistance)
-        {
-            return 1.0f;
-        }
-
-        var tangentDots = _particleData[particleIndex].Tangent.Dot(_particleData[particleIndex + 1].Tangent);
-
-        return
-            tangentDots >= Cos5Deg ? 1.0f :
-            tangentDots >= Cos15Deg ? 0.5f :
-            tangentDots >= Cos30Deg ? 0.33333f :
-            0.25f;
-    }
-
-    private void CalculateRopeCameraOrientation()
-    {
-        var cameraPosition = _camera?.GlobalPosition ?? Vector3.Zero;
-
-        ref var start = ref _particleData[0];
-        start.Tangent = (_particleData[1].PositionCurrent - start.PositionCurrent).Normalized();
-        start.Normal = (start.PositionCurrent - cameraPosition).Normalized();
-        start.Binormal = start.Normal.Cross(start.Tangent).Normalized();
-
-        ref var end = ref _particleData[SimulationParticles - 1];
-        end.Tangent = (end.PositionCurrent - _particleData[SimulationParticles - 2].PositionCurrent).Normalized();
-        end.Normal = (end.PositionCurrent - cameraPosition).Normalized();
-        end.Binormal = end.Normal.Cross(end.Tangent).Normalized();
-
-        for (var i = 1; i < SimulationParticles - 1; i++)
-        {
-            ref var particle = ref _particleData[i];
-            particle.Tangent = (_particleData[i + 1].PositionCurrent - _particleData[i - 1].PositionCurrent).Normalized();
-            particle.Normal = (_particleData[i].PositionCurrent - cameraPosition).Normalized();
-            particle.Binormal = _particleData[i].Normal.Cross(_particleData[i].Tangent).Normalized();
-        }
     }
 
     #endregion
@@ -410,7 +253,7 @@ public partial class VerletRope : MeshInstance3D
 
         if (visuals.Size == Vector3.Zero)
         {
-            return (false, Array.Empty<Vector3>());
+            return (false, []);
         }
 
         _collisionCheckBox.Size = visuals.Size;
@@ -433,7 +276,7 @@ public partial class VerletRope : MeshInstance3D
                 .ToArray();
         }
 
-        return (isStaticCollision, dynamicCollisions ?? Array.Empty<Vector3>());
+        return (isStaticCollision, dynamicCollisions ?? []);
     }
 
     private void CollideRope(ICollection<Vector3> dynamicCollisions)
@@ -501,48 +344,6 @@ public partial class VerletRope : MeshInstance3D
     }
 
     #endregion
-
-    private void DrawCurve()
-    {
-        // Catmull curve
-        _mesh.SurfaceBegin(Mesh.PrimitiveType.Triangles);
-
-        var cameraPosition = _camera?.GlobalPosition ?? Vector3.Zero;
-
-        for (var i = 0; i < SimulationParticles - 1; i++)
-        {
-            var (p0, p1, p2, p3) = GetSimulationParticles(i);
-            var step = GetDrawSubdivisionStep(cameraPosition, i);
-            var t = 0.0f;
-
-            while (t <= 1.0f)
-            {
-                CatmullInterpolate(p0, p1, p2, p3, 0.0f, t, out var currentPosition, out var currentTangent);
-                CatmullInterpolate(p0, p1, p2, p3, 0.0f, Mathf.Min(t + step, 1.0f), out var nextPosition, out var nextTangent);
-
-                var currentNormal = (currentPosition - cameraPosition).Normalized();
-                var currentBinormal = currentNormal.Cross(currentTangent).Normalized();
-                currentPosition -= GlobalPosition;
-
-                var nextNormal = (nextPosition - cameraPosition).Normalized();
-                var nextBinormal = nextNormal.Cross(nextTangent).Normalized();
-                nextPosition -= GlobalPosition;
-
-                var vs = new[]
-                {
-                    currentPosition - (currentBinormal * RopeWidth),
-                    nextPosition - (nextBinormal * RopeWidth),
-                    nextPosition + (nextBinormal * RopeWidth),
-                    currentPosition + (currentBinormal * RopeWidth)
-                };
-
-                DrawQuad(vs, -currentBinormal, t, t + step);
-                t += step;
-            }
-        }
-
-        _mesh.SurfaceEnd();
-    }
 
     private void VerletProcess(float delta)
     {
@@ -616,25 +417,14 @@ public partial class VerletRope : MeshInstance3D
 
     #endregion
 
-    public override string[] _GetConfigurationWarnings()
-    {
-        return UseVisibleOnScreenNotifier ? Array.Empty<string>() : new[] { NoNotifierWarning };
-    }
-
     public override void _Ready()
     {
+        base._Ready();
+
         if (!Engine.IsEditorHint() && StartDrawSimulationOnStart)
         {
             Draw = true;
             Simulate = true;
-        }
-
-        _mesh = Mesh as ImmediateMesh;
-        if (_mesh == null || _mesh.GetMeta(CreationStampMeta, 0ul).AsUInt64() != GetInstanceId())
-        {
-            Mesh = _mesh = new ImmediateMesh();
-            _mesh.SetMeta(CreationStampMeta, GetInstanceId());
-            _mesh.ResourceLocalToScene = true;
         }
 
         AddChild(_rayCast = new RayCast3D
@@ -645,14 +435,6 @@ public partial class VerletRope : MeshInstance3D
             Enabled = false
         });
 
-        if (UseVisibleOnScreenNotifier)
-        {
-            AddChild(_visibleNotifier = new VisibleOnScreenNotifier3D());
-            _visibleNotifier.ScreenEntered += () => Draw = true;
-            _visibleNotifier.ScreenExited += () => Draw = false;
-        }
-
-        _camera = GetViewport().GetCamera3D();
         _spaceState = GetWorld3D().DirectSpaceState;
         var visuals = GetAabb();
 
@@ -669,7 +451,6 @@ public partial class VerletRope : MeshInstance3D
         };
 
         _collisionShapeParameters.Transform = new Transform3D(_collisionShapeParameters.Transform.Basis, GlobalPosition + visuals.Position + (visuals.Size / 2));
-        MaterialOverride ??= GD.Load<StandardMaterial3D>(DefaultMaterialPath);
 
         CreateRope();
     }
@@ -684,10 +465,13 @@ public partial class VerletRope : MeshInstance3D
         _time += delta;
         _simulationDelta += delta;
 
-        var simulationStep = 1 / (float) SimulationRate;
-        if (_simulationDelta < simulationStep)
+        if (SimulationRate != 0)
         {
-            return;
+            var simulationStep = 1f / SimulationRate;
+            if (_simulationDelta < simulationStep)
+            {
+                return;
+            }
         }
 
         if (_attachEnd != null)
@@ -711,17 +495,12 @@ public partial class VerletRope : MeshInstance3D
 
         if (Draw)
         {
-            _camera = GetViewport().GetCamera3D();
-            CalculateRopeCameraOrientation();
-            _mesh.ClearSurfaces();
-            ResetRopeRotation();
-            DrawRopeDebugParticles();
-            DrawCurve();
+            DrawRopeParticles(_particleData);
+        }
 
-            if (_visibleNotifier != null)
-            {
-                _visibleNotifier.Aabb = GetAabb();
-            }
+        if (DrawDebugParticles)
+        {
+            DrawRopeDebugParticles(_particleData);
         }
 
         EmitSignal(SignalName.SimulationStep, _simulationDelta);
@@ -758,42 +537,11 @@ public partial class VerletRope : MeshInstance3D
             VerletProcess(PreprocessDelta);
             ApplyConstraints();
         }
-
-        CalculateRopeCameraOrientation();
     }
 
     public void DestroyRope()
     {
         _particleData.Resize(0);
         SimulationParticles = 0;
-    }
-
-    public void DrawRopeDebugParticles()
-    {
-        const float debugParticleLength = 0.3f;
-
-        if (!DrawDebugParticles)
-        {
-            return;
-        }
-
-        _mesh.SurfaceBegin(Mesh.PrimitiveType.Lines);
-
-        for (var i = 0; i < SimulationParticles; i++)
-        {
-            var particle = _particleData[i];
-            var localPosition = particle.PositionCurrent - GlobalPosition;
-
-            _mesh.SurfaceAddVertex(localPosition);
-            _mesh.SurfaceAddVertex(localPosition + (debugParticleLength * particle.Tangent));
-
-            _mesh.SurfaceAddVertex(localPosition);
-            _mesh.SurfaceAddVertex(localPosition + (debugParticleLength * particle.Normal));
-
-            _mesh.SurfaceAddVertex(localPosition);
-            _mesh.SurfaceAddVertex(localPosition + (debugParticleLength * particle.Binormal));
-        }
-
-        _mesh.SurfaceEnd();
     }
 }
