+++ addons/verlet_rope_4/structure/RopeCollisionType.cs
@@ -1,4 +1,4 @@
-﻿namespace VerletRope.addons.verlet_rope_4.Structure;
+﻿namespace VerletRope4.Structure;
 
 // ReSharper disable UnusedMember.Global
 public enum RopeCollisionType
