+++ demo/demo-legacy.tscn
@@ -1,6 +1,6 @@
-[gd_scene load_steps=10 format=3 uid="uid://bota3ppynjsu1"]
+[gd_scene load_steps=16 format=3 uid="uid://d0647cepiojoa"]
 
-[ext_resource type="Material" uid="uid://o7na8drc2je2" path="res://addons/verlet_rope_4/materials/rope_default.material" id="2_dar0o"]
+[ext_resource type="Material" uid="uid://o7na8drc2je2" path="res://addons/verlet_rope_4/materials/rope_default.material" id="1_xy6g0"]
 [ext_resource type="Script" uid="uid://v22nt3dj3ewl" path="res://addons/verlet_rope_4/VerletRope.cs" id="2_lows4"]
 [ext_resource type="Script" uid="uid://d36wx6t8jmjcx" path="res://demo/MovingRopeDemo.cs" id="2_vutyd"]
 
@@ -25,7 +25,32 @@ albedo_color = Color(0.529412, 0.12549, 0.133333, 1)
 
 [sub_resource type="ImmediateMesh" id="ImmediateMesh_dar0o"]
 resource_local_to_scene = true
-metadata/verlet_rope_internal_stamp = 14658790576662
+metadata/creation_stamp = 2201053379971
+
+[sub_resource type="FastNoiseLite" id="FastNoiseLite_0fwmc"]
+
+[sub_resource type="StandardMaterial3D" id="StandardMaterial3D_bbaqh"]
+cull_mode = 2
+albedo_color = Color(0.827451, 0.639216, 1, 1)
+disable_receive_shadows = true
+
+[sub_resource type="ImmediateMesh" id="ImmediateMesh_j5lb4"]
+resource_local_to_scene = true
+metadata/creation_stamp = 2201103711621
+
+[sub_resource type="ImmediateMesh" id="ImmediateMesh_u10mr"]
+resource_local_to_scene = true
+metadata/creation_stamp = 2201170820488
+
+[sub_resource type="StandardMaterial3D" id="StandardMaterial3D_8qm37"]
+cull_mode = 2
+vertex_color_use_as_albedo = true
+albedo_color = Color(0.32549, 0.407843, 0.219608, 1)
+disable_receive_shadows = true
+
+[sub_resource type="ImmediateMesh" id="ImmediateMesh_onu7k"]
+resource_local_to_scene = true
+metadata/creation_stamp = 2201221152139
 
 [node name="Demo" type="Node3D"]
 
@@ -56,8 +81,40 @@ material_override = SubResource("StandardMaterial3D_pbhg3")
 use_collision = true
 size = Vector3(0.2, 3, 0.2)
 
+[node name="VerletRope" type="MeshInstance3D" parent="PillarRope/Pillar" node_paths=PackedStringArray("AttachEnd")]
+transform = Transform3D(0.707107, 0, -0.707107, 0, 1, 0, 0.707107, 0, 0.707107, -0.0962079, 1.50411, 0.00534463)
+material_override = ExtResource("1_xy6g0")
+mesh = SubResource("ImmediateMesh_dar0o")
+script = ExtResource("2_lows4")
+AttachEnd = NodePath("../../Pillar2/PillarEnd")
+RopeLength = 3.0
+RopeWidth = 0.04
+PreprocessIterations = 0
+Simulate = false
+ApplyWind = true
+WindNoise = SubResource("FastNoiseLite_0fwmc")
+Wind = Vector3(1, 0, 1)
+WindScale = 30.0
+SlideIgnoreCollisionStretch = 1.25
+
 [node name="PillarEnd" type="Node3D" parent="PillarRope/Pillar"]
 transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0.00181818, 1.48938, 0.097929)
+
+[node name="VerletRope" type="MeshInstance3D" parent="PillarRope/Pillar/PillarEnd"]
+transform = Transform3D(0.707107, 0, -0.707107, 0, 1, 0, 0.707107, 0, 0.707107, 0, 0, 0)
+material_override = SubResource("StandardMaterial3D_bbaqh")
+mesh = SubResource("ImmediateMesh_j5lb4")
+script = ExtResource("2_lows4")
+RopeLength = 2.0
+SimulationParticles = 12
+PreprocessIterations = 0
+Simulate = false
+ApplyWind = true
+WindNoise = SubResource("FastNoiseLite_0fwmc")
+Wind = Vector3(1, 0, 1)
+WindScale = 30.0
+DampingFactor = 150.0
+SlideIgnoreCollisionStretch = 1.25
 
 [node name="Pillar2" type="CSGBox3D" parent="PillarRope"]
 transform = Transform3D(0.707107, 0, 0.707107, 0, 1, 0, -0.707107, 0, 0.707107, -0.937059, 1.4665, 1.02452)
@@ -68,8 +125,25 @@ size = Vector3(0.2, 3, 0.2)
 [node name="PillarEnd" type="Node3D" parent="PillarRope/Pillar2"]
 transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0.102865, 1.48938, -0.00311756)
 
-[node name="MovingRope" type="Node3D" parent="."]
+[node name="VerletRope" type="MeshInstance3D" parent="PillarRope/Pillar2/PillarEnd"]
+transform = Transform3D(0.707107, 0, -0.707107, 0, 1, 0, 0.707107, 0, 0.707107, -0.0998901, 0, 0.0998905)
+material_override = SubResource("StandardMaterial3D_bbaqh")
+mesh = SubResource("ImmediateMesh_u10mr")
+script = ExtResource("2_lows4")
+RopeLength = 2.0
+SimulationParticles = 12
+PreprocessIterations = 0
+Simulate = false
+ApplyWind = true
+WindNoise = SubResource("FastNoiseLite_0fwmc")
+Wind = Vector3(1, 0, 1)
+WindScale = 30.0
+DampingFactor = 150.0
+SlideIgnoreCollisionStretch = 1.25
+
+[node name="MovingRope" type="Node3D" parent="." node_paths=PackedStringArray("MovingNode")]
 script = ExtResource("2_vutyd")
+MovingNode = NodePath("VerletRopeMoving")
 MovingPath = Vector3(7, 0, 0)
 PathTime = 7.0
 TransitionType = 7
@@ -81,6 +155,23 @@ use_collision = true
 radius = 1.46126
 height = 3.10616
 sides = 50
+
+[node name="VerletRopeMoving" type="MeshInstance3D" parent="MovingRope"]
+transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0.537658, 2.69336, 5.85259)
+material_override = SubResource("StandardMaterial3D_8qm37")
+mesh = SubResource("ImmediateMesh_onu7k")
+skeleton = NodePath("../..")
+script = ExtResource("2_lows4")
+RopeLength = 3.0
+SimulationParticles = 15
+SimulationRate = 265
+WindNoise = SubResource("FastNoiseLite_0fwmc")
+Wind = Vector3(1, 0, 1)
+WindScale = 10.0
+DampingFactor = 1000.0
+RopeCollisionType = 2
+RopeCollisionBehavior = 2
+DynamicCollisionMask = 2
 
 [node name="MovingCylinder" type="CSGCylinder3D" parent="MovingRope" node_paths=PackedStringArray("MovingNode")]
 transform = Transform3D(0.999993, 0.00371743, -1.62487e-10, -3.88745e-16, -4.37114e-08, -1, -0.00371743, 0.999993, -4.37111e-08, -0.297686, 1.49849, 5.83115)
@@ -94,12 +185,3 @@ MovingNode = NodePath(".")
 MovingPath = Vector3(3, 0, 0)
 PathTime = 5.0
 TransitionType = 1
-
-[node name="VerletRope" type="MeshInstance3D" parent="MovingRope"]
-transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 4.85672, 6.56651, 4.66125)
-material_override = ExtResource("2_dar0o")
-mesh = SubResource("ImmediateMesh_dar0o")
-script = ExtResource("2_lows4")
-RopeCollisionBehavior = 2
-RopeLength = 3.0
-metadata/_custom_type_script = "uid://v22nt3dj3ewl"
